+++ Makefile
@@ -18,12 +18,12 @@ ENVIRON 	=	src/environs.c \
 
 OPERATORS 	=	src/operators.c \
 
-PARSER		=	src/luis_redirections.c \
-				src/luis_commands_utils.c \
-				src/luis_commands.c \
-				src/luis_org_tok.c \
-				src/luis_ast_create.c \
-				src/parsing.c \
+PARSER		=	src/parsing.c \
+				# src/luis_redirections.c \
+				# src/luis_commands_utils.c \
+				# src/luis_commands.c \
+				# src/luis_org_tok.c \
+				# src/luis_ast_create.c \
 
 BUILTIN 	=	src/built-ins/ft_env.c \
 				src/built-ins/ft_pwd.c \
