+++ include/builtins.h
@@ -1,0 +1,21 @@
+#ifndef BUILTINS_H
+# define BUILTINS_H
+
+# include "minishell.h"  
+# include "../libs/libft/libft.h"
+
+// built-ins
+void	ft_env(t_list *envlist);
+void	ft_pwd(t_list **envlist);
+void	ft_echo(char *line, bool flag);
+void	ft_cd(t_list **envlist, char *new_path);
+void	ft_exit(void);
+void	ft_export(t_list **envlist, char *variable_value);
+void	ft_unset(t_list **envlist, char *variable);
+
+// built-ins helpers
+void	ft_ms_env_add(t_list **envlist, char *variable_value);
+void	ft_ms_env_update_export(t_list **envlist, char *variable, char *value);
+void	ft_ms_env_update_cd(t_list **envlist, char *variable, char *value);
+
+#endif
