+++ src/built-ins/ft_cd.c
@@ -10,7 +10,13 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/minishell.h"
+#include "../../include/minishell.h"  
+#include "../../include/heredoc.h"
+#include "../../include/expand.h"
+#include "../../include/tokenize.h"
+#include "../../include/environs.h"
+#include "../../include/heredoc.h"
+#include "../../include/builtins.h"
 
 // char	*ft_cd_path_treatment(char *ms_env[], char *path)
 // {
