+++ src/heredocs.c
@@ -11,25 +11,17 @@
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
-
-char	*ft_hd_create_file(int *hd_count_int, char **filepath);
-char	*ft_hd_input_loop(t_list **envlist, t_mem **mem);
-void	*ft_hd_write_to_file(int hd_loop_count, t_mem **mem);
-char	*ft_hd_validate_path(char **filepath, int *hd_count_int);
-int		ft_hd_init_file(char **filepath);
-void	ft_del_heredoc_node(void *content);
-void	reset(t_mem **mem);
-
+#include "../include/heredoc.h"
+#include "../include/expand.h"
 
 char	*ft_heredoc(char *delimiter, t_mem **mem)
 {
 	static int	hd_count_int;
-	t_hd_mem	*hd;
+	t_hdc_mem	*hd;
 	t_env_mem	*env;
 
 	hd = (*mem)->heredoc;
 	env = (*mem)->environs;
-
 
 	hd->delim = ft_strdup(delimiter);
 	ft_expand(&hd->delim, DELIMITER, mem);
@@ -69,7 +61,7 @@ char	*ft_hd_create_file(int *hd_count_int, char **filepath)
 void	*ft_hd_write_to_file(int hd_loop_count, t_mem **mem)
 {
 	int			hd_temp_file;
-	t_hd_mem	*hd; 
+	t_hdc_mem	*hd; 
 
 	hd = (*mem)->heredoc;
 
@@ -97,7 +89,7 @@ void	*ft_hd_write_to_file(int hd_loop_count, t_mem **mem)
 char	*ft_hd_input_loop(t_list **envlist, t_mem **mem)
 {
 	int			hd_loop_count;
-	t_hd_mem	*hd; 
+	t_hdc_mem	*hd; 
 
 	hd =(*mem)->heredoc;
 	(void)envlist;
@@ -173,24 +165,23 @@ int	ft_hd_init_file(char **filepath)
 	return (0);
 }
 
-void	ft_del_heredoc_node(void *content)
+
+void	*ft_init_hdc_memory(t_mem **mem)
 {
-	t_hd_node	*hd_node;
-
-	if (!content)
-		return ;
-	
-	hd_node = (t_hd_node *)content;
-	if (!hd_node->fpath)
-		return ;
-
-	if (access(hd_node->fpath, F_OK) == 0)
-	{
-		unlink(hd_node->fpath);
-		ft_free_and_null((void *)&hd_node->fpath);
-	}
-	ft_free_and_null((void *)&hd_node);
+	(*mem)->heredoc = malloc(sizeof(t_hdc_mem));
+	if(!(*mem)->heredoc)
+		return (NULL);
+	(*mem)->heredoc->delim = NULL;
+	(*mem)->heredoc->filepath = NULL;
+	(*mem)->heredoc->looptemp = NULL;
+	(*mem)->heredoc->loopinput = NULL;
+	(*mem)->heredoc->mode = INIT_MODE;
+	return ((*mem)->heredoc);
 }
 
-
-
+void	ft_clear_hdc_mem(t_hdc_mem **hd)
+{
+	ft_free_and_null((void *)&(*hd)->delim);
+	free(*hd);
+	return ;
+}
